fix(build): fix build breakages introduced by new data grid
new data grid brought *lots* of added complexity to the build
process, so much so that node was running out of ram during the
build-renderer phase. I've mitigated this by combing through
& upgrading *all* dependencies to the latest versions.

This required some code changes, namely jumping a major version
for normalizr (which seems like it'll come with it's own benefits).

I'd initially thought we'd have to up the memory allocation for
the node process that's building the renderer, but after upgrading
packages the issue cleared up.