feat(backend): Use PATH qri command & repo
Having a duplicate, internal copy of the qri CLI inside the app
is officially causing too many problems. Alter app/backend.js
to stop spinning up it's own backend & repo, replacing it with
a version check & call to start a canonical qri server.

This brings some future technical debt in the form of needing
to port this check/update process to non-unix systems, but is
the procedure we follow regardless of platform.

closes #209