fix(Dataset): load most recent version of dataset
When we hit ‘save’ on the MetadataEditor, we call saveMetaDataset,
which calls to the `save/peername/datasetname` qri API endpoint. This
returns an updated dataset, which gets stored in the state tree under
‘entities’
when we go to the Dataset page for a specific dataset, we use the
selectDatasetByName function to grab the dataset. However, when we
iterate through the datasets in ‘namespace’ we will always grab the
first one we hit. After we’ve saved the metadata, however, this new
dataset (the one we want to grab, because it is the most recent version
of the dataset), will never get selected because it comes after the
original dataset in the array. The quick fix is to reverse the array
before we look for the correct peername/datasetname combination, but
this feels like a hack and something we need to figure a better way to
select.