fix(AddDataset, Dropfile): fix bug that wasn't updating state
All our components extend the Base component, which in turn extends the
React.PureComponent.

However, we have a lot of changing/moving parts that still need to
update when the state changes. We need to make sure we are setting the
state in the right way to ensure the component gets updated.