chore(Webapp build): make webapp build output a single file
Ok, for the sake of simplicity we're currently going to have the webapp compile
down to a single file so we can ship it over the d.web as a self-contained app.

The problem here stems from not being able to get webpack fetches to happen in the
context of a directory inside a hash-path. We either need to proxy requests for certian
files to a hash-directory on the webapp server in qri-io/qri, or find some way to
mess with webpack's runtime "import" calls

In the future we should *absolutely* imporove on this situation, but for now this works.