feat(readOnly): add ReadOnlyBlock and ReadOnlyInline components
these components are displayed when the app is in read only mode. ReadOnly mode is set in the webpack.config.base.js file, in the plugins section. This is probably not the best way to do this, but it is what it is for now.

READ_ONLY should be set to a bool, READ_ONLY_LINKS is an array of objects, with the keys "name" and "link" (to be displayed in the ReadOnlyInline component), and READ_ONLY_TEXT is a string (to be displayed in the ReadOnlyBlock component)