fix(AppDrag): added back `Header` component, renamed to `AppDrag`
Needed so that the app can be moved in electron mode. There is an electron bug, however, that does not allow you to move the app when chrome DevTools is docked to the left, bottom, or right. https://github.com/electron/electron/issues/3647 for more info, but for now, you can either open the inspector in a seperate windown, or close the inspector when you need to move the app