refactor(dataset): dataset now shows correct options for editing etc
metadata editor and rename buttons are shown:
if !peer
if latestDataset
if sessionProfileName === peername

remove button is shown if:
if !peer
if sessionProfileName === peername

add button is shown if:
if peer

export button and changes tab are shown if:
if !peer

This commit also moves all the logic from inside the `handleDeleteDataset` functions (and other similar functions), to where we pass those functions in in the first place. Moves the logic for pulling out the information we need from props to each of those functions.