refactor(dataset): add DatasetSummary as default route
Add `datasetContent` class to scss, that way if we want to change the layout in the future, we only need to change it in one place

In DatasetNav, changed  `Link` to `NavLink`, because `NavLink` is supposed to allow you to add style to active links. This seems to be not working right now, boo.

Added DatasetSummary, which is basically the old Meta component.
Added `/summary` as a nested route in `Dataset`