refactor(Header, DatasetNav): generalize components for re-use
Generalize Header and DatasetNav:

DatasetNav renamed to NavLinks, for more general use.
NavLinks takes a list of names & links, a url, and a bool indicating if we want large or small links (small links are slightly smaller and blue)

Header takes a left child, right child, and a url and list of links for a NavLinks component.

DatasetHeader now wraps Header.