refactor(datasets, search, topbar): search takes you to datasets search results page
New nested routes in Datasets page: `/local` and `/search`, which take you to your datasets and your search results respectively. The goal is that the last searched term and datasets will always be displayed in the search results page, but this is currently not wired up.

ViewMode is removed from the TopBar component