refactor(TopBar, SearchResults): separate location from search results
Previous, we were conflating search results with the text that gets typed in the nav bar, that is, we were only keeping track of the text in the nav bar and not separately keeping track of our previous search string.

This commit separates the two. The `app` section of the state tree now contains a `location` and `search` field.

Updated app reducer tests