docs(COMMON_PITFALLS): add a doc where we can document head scratching bugs and how we solved them
I've found I have developed a bit of a reference internally of what to look for when a weird bug crops us. The COMMON_PITFALLS doc is a place where I can document those places to look. It will grow as I remember things and in the future as I bang my head against the wall looking for a bug only to find it was caused by one dumb thing I did one time and then have to ask Brendan for help fixing it.