refactor(Transform, Viz): quick first attempt to at transform and viz
Right now, the Transform and Viz components wrap an iframe, that pulls the transform and viz data right off the d-web.

The viz data, however, gets interpreted as html, so we can't display it as plain text.

I'm going to commit this for now, and start a new PR that will attempt to write middleware to begin to address this problem.