fix(Editor): body is fully removed when you hit the remove button
When a dataset comes in with a body already defined, not only lives in the state tree under the `editor`, but it also lives in the state tree under `editor.dataset`.

In order to ensure that no traces of the body remain once we remove,  we extended the `removeSection` function to also be able to remove all elements of the body from the `editor.dataset`. Use this function to remove the body when you hit the remove button. Also use the `onSetBody` function to remove the body from the state tree.

Let's also note that the body is existing in two sections and that perhaps we should just let it live in the `editor.dataset`