refactor(state.editor): have two seperate sections for stringified and parsed data
the dream here is to have two seperate sections for our stringified data, that we need for the monico editor and our parsed data, that would all live in dataset (eventually, for now body stays in the` state.editor.body`). Adds `stringified`section to `state.editor`
One day perhaps it will hold all of:
    - bodyString
    - schemaString
    - vizScript
    - transformScript

And any other part of the dataset that can be manually edited and isn't immediately in the state tree as a well defined section of a dataset (the way that metadata is)

This is the structure:
```
  {
    editor:
      stringified: {
        schema: {
          string: "",
          parsable: true
        },
        body: {
          string: "",
          parsable: true
        },
        etc...
      }
    }
  }
```

  if we get code completion and some more sophisitcated editing capabilites
  maybe we can figure out a way to know if the viz or transform
  are well formed code

We will use this "parsable" section to determine if we can send the code over and what we need the user to fix before we are able to send it over.

This commit also rewires where we grab the schema and body from, in their respective editor sections.