feat(EditBody): move, add, and delete columns and rows & data is persistent
Yay! now any changes the user makes to the table, aka moves a row or column, adds a row or column, the changes are preserved when the user moves away from the body editor and back again!

Right now the `rowOrder` and `colOrder` which are tracking the row and column switches are in the `Editor` component state. They should be moved to the state tree.

There are definitely many candidates for refactoring from this pr, alot of the code can be made more concise/reused.

I am super pumped about this functionality and how these changes are an incremental step toward tracking _changes_ on the editor rather than the full state.