refactor(rowOrder, colOrder): move row and col order to state tree
To prep for removing `stringified`, need to first move rowOrder and colOrder into the state tree.

Also, moved the `onCollapseChanges` from the `Editor` component. Instead of having access to props or state, it is now a factory that takes a body, columnHeaders (`schema.items.items`), rowOrder, and colOrder, and returns a `changed` object, made up of two keys: body and schema.

If there is no rowOrder, or colOrder, or body, or columnHeaders, it returns undefined. If `changed` is undefined, we know not to try to set the body or schema.

`onCollapseChanges` has been renamed to `generateCollapsedChanges` and has been moved to a util file. This util file has been renamed from `generateSchema` to `generate`, not married to the name. Just want a seperate file of little machines that create new versions/new objects.

We need this `generateCollapsedChanges` in many places in our editor codepath, so instead of writing it over, figured it'd be good to have a function handy