refactor(parsable): remove all references to `parsable`
Removed references to `bodyParsable`, `parsable`, and `selectDatasetValid`, which all rely on `stringified.foo.parsable`

This also means we needed to write another little machine function `generateValid` that takes a body and structure (and colOrder and rowOrder). It returns an error if the schema or body cannot be parsed, and a `valid` object with keys `body` and `structure` if it can be. (replaces `selectDatasetValid`)

Also, took better notes/left comments on `onSetView`, so it will be easier in the next pass when we switch over to `onSetView` as a function on the reducer and not a local `EditBody` function.

Removed references to handling csv for now, as all body going over the frontend is json. This will be added back in when you can upload a file.