feat(Upload Body): initial steps upload button on edit body page
when you click the label, you can upload a file. If there are multiple files selected or the files are not in the correct format, there is an alert explaining.

If it is in one of the three welcomed formats (json, csv, cbor), we make sure to store this data.

This is where I ran into some hiccups and so have left this a little raw while we proceed.

We want a user to be able to upload a body, then navigate away from the editor and back and the file still be uploaded. So, in that case, we need to store the file in the state tree. But storing non-serializable stuff in the state tree is frowned upon. So instead, storing it in base64, and will pull it back out of base64 when we go to save.