fix(dataset): add silent error to dataset and registry load dataset actions
Because we are loading from two sources, only one source needs to load correctly to give the user a good experience.

However, right now, if one source fails, it throws an error that is visible to the user, even though they can see their dataset is correctly loaded.

To better control what the user is going to see, let's make these errors silent