fix(dataset): add logic to dataset page to show correct error messaging
We need to compare the `action.error` that returns from the local node and the registry. If we errors back from both locations, then show an error state to the user. If we don't get errors back from both, and we previously were showing an errored state, hide that error.

There is definitely a better way to chain these errors more cleanly, but leaving this for now.