refactor(AppLoading): move css to inline
The AppLoading component renders before the app is fully loaded! So the css isn't fully loaded when the component first renders. It loads after a split second, injecting the css into the component and re-rendering. However, it causes an ugly shutter in our app.

Moving the css to inline for this component to it renders correctly on the first go round!