refactor(TopBar, Routes): use build variable READONLY to code split!
First, in order to use the dynamic import syntax, we need to include the babel dynamic-import plugin.

Then, let's use the build var READONLY to only show a link to the editor when READONLY is false.

Finally, let's do the actual code splitting. We load the `EditorContainer` using React.lazy, and wrap the <Route> in React.Suspense. Our fallback is the spinner we usually use for loading.

tada!