feat(*): add pagination to the body
This pr adds pagination to the body.

To keep it simple, for now, we are using a `LoadMore` component that has 3 buttons, each adding 100, 500, or 1000 rows respectfully.

when you get to the bottom of a dataset (tabular or not) that has not yet `fetchedAll`, the `LoadMore` component will show.

Body data has been moved out of pagination, and out of entities (more on that later) and into it's own `body` section of the state tree.

The body section contains an `id` (the bodypath), `selector` (stub for the future when we want to filter results), `fetchedAll`, `pageSize`, `error`, and `data` (the actual body data).

We've added a new body reducer and some body actions!

The reducer has 4 options:
- CLEAR_BODY, which returns the initial state
- BODY_REQUEST, which, if the bodypaths or the selectors don't match, returns an inital state with the new bodypath, selector, and pageSize, or if it is the same bodypath and selector, updates the state tree if the pageSize has changed. Either way, `loading` is set to true
- BODY_SUCCESS, which calculates if we have `fetchedAll`, appends the new  body data to any current body data, and sets `loading` to false
- BODY_FAILURE, which sets the `error` and sets `loading` to false

There are tests covering the body reducer

Some actions and selectors have changed!

The `selectDatasetBody` selector now checks to see if the `bodypath` given is the `bodypath` in `state.body.id`, before returning the body.

The actions `loadDatasetBody` and `fetchDatasetBody` actions have been moved to the `actions/body.js` file, and renamed to `loadBodyByPage`, and `fetchBodyByPage`.

We've added `loadBody` and `fetchBody`, that use `offset` and `limit`. This is the main action we use to load the body. We also have `clearBody`, which resets the `state.body`, and we call whenever we navigate away from the dataset page.

In order to use the callApi in our actions, but not have body end up in the `state.entities` tree, we wrote a very light piece of middleware that sits after the callApi, and callCaf middleware that moves the body data out of band if the `action.type` is `BODY_SUCCESS`.