feat(DownloadBar): add `DownloadBar` component
DownloadBar will be shown on the dataset page when in readonly mode. The DownloadBar will directy you to download qri if you want to see the underlying data, and link to our docs for guidance. The plan is to make a quick video tutorial that shows someone how to go from the `app.qri.io` page to getting the data on their machine